requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-docs
python-docx
scikit-learn
textstat
spacy
sentence-transformers
numpy
requests
pymupdf
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
google-generativeai
chromadb
python-dotenv
pandas
streamlit
google-genai
supabase